+++ OpenSim/Simulation/Model/GeometryPath.h
@@ -25,8 +25,7 @@
 
 
 // INCLUDE
-#include <OpenSim/Simulation/osimSimulationDLL.h>
-#include "OpenSim/Simulation/Model/ModelComponent.h"
+#include "OpenSim/Simulation/Model/AbstractPath.h"
 #include "PathPointSet.h"
 #include <OpenSim/Simulation/Wrap/PathWrapSet.h>
 #include <OpenSim/Simulation/MomentArmSolver.h>
@@ -55,32 +54,18 @@ class WrapObject;
  * @author Peter Loan
  * @version 1.0
  */
-class OSIMSIMULATION_API GeometryPath : public ModelComponent {
-OpenSim_DECLARE_CONCRETE_OBJECT(GeometryPath, ModelComponent);
-    //=============================================================================
-    // OUTPUTS
-    //=============================================================================
-    OpenSim_DECLARE_OUTPUT(length, double, getLength, SimTK::Stage::Position);
-    // 
-    OpenSim_DECLARE_OUTPUT(lengthening_speed, double, getLengtheningSpeed,
-        SimTK::Stage::Velocity);
+class OSIMSIMULATION_API GeometryPath : public AbstractPath {
+OpenSim_DECLARE_CONCRETE_OBJECT(GeometryPath, AbstractPath);
 
 //=============================================================================
 // DATA
 //=============================================================================
-public:
-    OpenSim_DECLARE_UNNAMED_PROPERTY(Appearance,
-        "Default appearance attributes for this GeometryPath");
-
 private:
     OpenSim_DECLARE_UNNAMED_PROPERTY(PathPointSet,
         "The set of points defining the path");
 
     OpenSim_DECLARE_UNNAMED_PROPERTY(PathWrapSet,
         "The wrap objects that are associated with this path");
-
-    // used for scaling tendon and fiber lengths
-    double _preScaleLength;
 
     // Solver used to compute moment-arms. The GeometryPath owns this object,
     // but we cannot simply use a unique_ptr because we want the pointer to be
@@ -135,18 +120,6 @@ public:
     // GET
     //--------------------------------------------------------------------------
 
-    /** If you call this prior to extendAddToSystem() it will be used to initialize
-    the color cache variable. Otherwise %GeometryPath will choose its own
-    default which varies depending on owner. **/
-    void setDefaultColor(const SimTK::Vec3& color) {
-        updProperty_Appearance().setValueIsDefault(false);
-        upd_Appearance().set_color(color); 
-    };
-    /** Returns the color that will be used to initialize the color cache
-    at the next extendAddToSystem() call. The actual color used to draw the path
-    will be taken from the cache variable, so may have changed. **/
-    const SimTK::Vec3& getDefaultColor() const { return get_Appearance().get_color(); }
-
     /** %Set the value of the color cache variable owned by this %GeometryPath
     object, in the cache of the given state. The value of this variable is used
     as the color when the path is drawn, which occurs with the state realized 
@@ -164,8 +137,6 @@ public:
 
     double getLength( const SimTK::State& s) const;
     void setLength( const SimTK::State& s, double length) const;
-    double getPreScaleLength( const SimTK::State& s) const;
-    void setPreScaleLength( const SimTK::State& s, double preScaleLength);
     const Array<AbstractPathPoint*>& getCurrentPath( const SimTK::State& s) const;
 
     double getLengtheningSpeed(const SimTK::State& s) const;
@@ -186,13 +157,15 @@ public:
     void addInEquivalentForces(const SimTK::State& state,
                                const double& tension, 
                                SimTK::Vector_<SimTK::SpatialVec>& bodyForces,
-                               SimTK::Vector& mobilityForces) const;
-
-
+                               SimTK::Vector& mobilityForces) const override;
+    
+    bool isVisualPath() const override { return true; }
+    
     //--------------------------------------------------------------------------
     // COMPUTATIONS
     //--------------------------------------------------------------------------
-    virtual double computeMomentArm(const SimTK::State& s, const Coordinate& aCoord) const;
+    double computeMomentArm(const SimTK::State& s,
+                            const Coordinate& aCoord) const override;
 
     //--------------------------------------------------------------------------
     // SCALING