+++ OpenSim/Simulation/Model/PathSpring.cpp
@@ -25,8 +25,6 @@
 // INCLUDES
 //=============================================================================
 #include "PathSpring.h"
-#include "GeometryPath.h"
-#include "PointForceDirection.h"
 
 //=============================================================================
 // STATICS
@@ -64,7 +62,7 @@ PathSpring::PathSpring(const string& name, double restLength,
 void PathSpring::constructProperties()
 {
     setAuthors("Ajay Seth");
-    constructProperty_GeometryPath(GeometryPath());
+    constructProperty_path(GeometryPath());
     constructProperty_resting_length(SimTK::NaN);
     constructProperty_stiffness(SimTK::NaN);
     constructProperty_dissipation(SimTK::NaN);
@@ -106,8 +104,7 @@ void PathSpring::extendFinalizeFromProperties()
 {
     Super::extendFinalizeFromProperties();
 
-    GeometryPath& path = upd_GeometryPath();
-    path.setDefaultColor(DefaultPathSpringColor);
+    updPath().setDefaultColor(DefaultPathSpringColor);
 
     OPENSIM_THROW_IF_FRMOBJ(
         (SimTK::isNaN(get_resting_length()) || get_resting_length() < 0),
@@ -139,7 +136,7 @@ void PathSpring::extendFinalizeFromProperties()
  */
 double PathSpring::getLength(const SimTK::State& s) const
 {
-    return getGeometryPath().getLength(s);
+    return getPath().getLength(s);
 }
 
 double PathSpring::getStretch(const SimTK::State& s) const
@@ -152,7 +149,7 @@ double PathSpring::getStretch(const SimTK::State& s) const
 
 double PathSpring::getLengtheningSpeed(const SimTK::State& s) const
 {
-    return getGeometryPath().getLengtheningSpeed(s);
+    return getPath().getLengtheningSpeed(s);
 }
 
 double PathSpring::getTension(const SimTK::State& s) const
@@ -174,13 +171,13 @@ extendPostScale(const SimTK::State& s, const ScaleSet& scaleSet)
 {
     Super::extendPostScale(s, scaleSet);
 
-    GeometryPath& path = upd_GeometryPath();
+    AbstractPath& path = updPath();
     if (path.getPreScaleLength(s) > 0.0)
     {
         double scaleFactor = path.getLength(s) / path.getPreScaleLength(s);
         upd_resting_length() *= scaleFactor;
 
-        // Clear the pre-scale length that was stored in the GeometryPath.
+        // Clear the pre-scale length that was stored in the AbstractPath.
         path.setPreScaleLength(s, 0.0);
     }
 }
@@ -191,9 +188,10 @@ extendPostScale(const SimTK::State& s, const ScaleSet& scaleSet)
 /**
  * Compute the moment-arm of this muscle about a coordinate.
  */
-double PathSpring::computeMomentArm(const SimTK::State& s, const Coordinate& aCoord) const
+double PathSpring::computeMomentArm(const SimTK::State& s,
+        const Coordinate& aCoord) const
 {
-    return getGeometryPath().computeMomentArm(s, aCoord);
+    return getPath().computeMomentArm(s, aCoord);
 }
 
 
@@ -202,17 +200,7 @@ void PathSpring::computeForce(const SimTK::State& s,
                               SimTK::Vector_<SimTK::SpatialVec>& bodyForces, 
                               SimTK::Vector& generalizedForces) const
 {
-    const GeometryPath& path = getGeometryPath();
+    const AbstractPath& path = getPath();
     const double& tension = getTension(s);
-
-    OpenSim::Array<PointForceDirection*> PFDs;
-    path.getPointForceDirections(s, &PFDs);
-
-    for (int i=0; i < PFDs.getSize(); i++) {
-        applyForceToPoint(s, PFDs[i]->frame(), PFDs[i]->point(), 
-                          tension*PFDs[i]->direction(), bodyForces);
-    }
-
-    for(int i=0; i < PFDs.getSize(); i++)
-        delete PFDs[i];
+    path.addInEquivalentForces(s, tension, bodyForces, generalizedForces);
 }